edge_ai_optimization.py
import jax
import jax.numpy as jnp
import flax
import optax
from flax import linen as nn
from typing import List, Tuple

class EdgeAIOptimization:
    def __init__(self):
        self.model = None
        self.optimizer = None

    def quantize_model(self, model: nn.Module, num_bits: int = 8) -> nn.Module:
        """
        Quantize the model to reduce its size and improve inference speed.
        """
        # Implement quantization logic here
        # This is a placeholder and needs to be implemented with actual quantization logic
        return model

    def prune_model(self, model: nn.Module, sparsity: float = 0.5) -> nn.Module:
        """
        Prune the model to reduce its size and computational requirements.
        """
        # Implement pruning logic here
        # This is a placeholder and needs to be implemented with actual pruning logic
        return model

    def knowledge_distillation(self, teacher_model: nn.Module, student_model: nn.Module,
                               data: jnp.ndarray) -> nn.Module:
        """
        Perform knowledge distillation to transfer knowledge from a larger model to a smaller one.
        """
        # Implement knowledge distillation logic here
        # This is a placeholder and needs to be implemented with actual distillation logic
        return student_model

    def setup_federated_learning(self, num_clients: int):
        """
        Set up a federated learning environment.
        """
        # Implement federated learning setup
        # This is a placeholder and needs to be implemented with actual federated learning logic
        pass

    def integrate_snn(self, input_size: int, num_neurons: int) -> nn.Module:
        """
        Integrate Spiking Neural Network (SNN) capabilities.
        """
        # Implement SNN integration
        # This is a placeholder and needs to be implemented with actual SNN logic
        class SNN(nn.Module):
            @nn.compact
            def __call__(self, x):
                # Placeholder for SNN logic
                return x
        return SNN()

    def self_healing_mechanism(self, model: nn.Module, data: jnp.ndarray) -> nn.Module:
        """
        Implement self-healing mechanisms using anomaly detection.
        """
        # Implement self-healing logic here
        # This is a placeholder and needs to be implemented with actual self-healing logic
        return model

    def optimize_latency_and_power(self, model: nn.Module) -> Tuple[nn.Module, float, float]:
        """
        Optimize the model for latency and power consumption.
        Returns the optimized model, estimated latency, and estimated power consumption.
        """
        # Implement optimization logic here
        # This is a placeholder and needs to be implemented with actual optimization logic
        return model, 0.0, 0.0

    def train_model(self, model: nn.Module, data: jnp.ndarray, labels: jnp.ndarray,
                    num_epochs: int = 10, learning_rate: float = 1e-3):
        """
        Train the model using the provided data and labels.
        """
        self.model = model
        self.optimizer = optax.adam(learning_rate)

        # Ensure labels are 1D
        labels = jnp.squeeze(labels)

        params = self.model.init(jax.random.PRNGKey(0), jnp.ones((1, *data.shape[1:])))
        opt_state = self.optimizer.init(params)

        @jax.jit
        def train_step(params, opt_state, batch, labels):
            def loss_fn(params):
                logits = self.model.apply(params, batch)
                logits = logits.reshape(logits.shape[0], -1)  # Reshape logits to 2D
                loss = optax.softmax_cross_entropy_with_integer_labels(logits, labels).mean()
                return loss

            loss, grads = jax.value_and_grad(loss_fn)(params)
            updates, new_opt_state = self.optimizer.update(grads, opt_state)
            new_params = optax.apply_updates(params, updates)
            return new_params, new_opt_state, loss

        for epoch in range(num_epochs):
            params, opt_state, loss = train_step(params, opt_state, data, labels)
            print(f"Epoch {epoch+1}, Loss: {loss}")

        self.model = self.model.bind(params)
        self.trained_params = params  # Store the trained parameters

    def infer(self, data: jnp.ndarray) -> jnp.ndarray:
        """
        Perform inference using the trained model.
        """
        if self.model is None or not hasattr(self, 'trained_params'):
            raise ValueError("Model has not been trained yet.")
        # Ensure input data is reshaped to (1, 28, 28, 1) for a single image
        data = data.reshape(1, 28, 28, 1)
        output = self.model.apply(self.trained_params, data)
        return output

# Example usage
if __name__ == "__main__":
    edge_ai = EdgeAIOptimization()

    # Create a simple model
    class SimpleModel(nn.Module):
        @nn.compact
        def __call__(self, x):
            print(f"Input shape: {x.shape}")
            x = nn.Conv(features=32, kernel_size=(3, 3))(x)
            print(f"After first Conv shape: {x.shape}")
            x = nn.relu(x)
            x = nn.avg_pool(x, window_shape=(2, 2), strides=(2, 2))
            print(f"After first avg_pool shape: {x.shape}")
            x = nn.Conv(features=64, kernel_size=(3, 3))(x)
            print(f"After second Conv shape: {x.shape}")
            x = nn.relu(x)
            x = nn.avg_pool(x, window_shape=(2, 2), strides=(2, 2))
            print(f"After second avg_pool shape: {x.shape}")
            x = x.reshape((x.shape[0], -1))  # Flatten the input
            print(f"After reshape shape: {x.shape}")
            x = nn.Dense(features=64)(x)
            print(f"After first Dense shape: {x.shape}")
            x = nn.relu(x)
            x = nn.Dense(features=10)(x)
            print(f"Final output shape: {x.shape}")
            return x  # Output is 2D (batch_size, num_classes)

    model = SimpleModel()

    # Generate some dummy data
    data = jax.random.normal(jax.random.PRNGKey(0), (100, 28, 28, 1))
    labels = jax.random.randint(jax.random.PRNGKey(1), (100,), 0, 10)

    # Train the model
    edge_ai.train_model(model, data, labels)

    # Optimize the model
    optimized_model = edge_ai.quantize_model(model)
    optimized_model = edge_ai.prune_model(optimized_model)

    # Perform inference
    result = edge_ai.infer(jax.random.normal(jax.random.PRNGKey(2), (1, 28, 28, 1)))
    print("Inference result:", result)

    # Note: The above example is simplified and doesn't fully implement all methods.
    # Each method needs to be properly implemented with the actual logic for edge AI optimization.
