test_edge_ai_optimization.py
import unittest
import jax
import jax.numpy as jnp
from edge_ai_optimization import EdgeAIOptimization
from flax import linen as nn

class TestEdgeAIOptimization(unittest.TestCase):
    def setUp(self):
        self.edge_ai = EdgeAIOptimization()

        class SimpleModel(nn.Module):
            @nn.compact
            def __call__(self, x):
                x = nn.Conv(features=32, kernel_size=(3, 3))(x)
                x = nn.relu(x)
                x = nn.avg_pool(x, window_shape=(2, 2), strides=(2, 2))
                x = nn.Conv(features=64, kernel_size=(3, 3))(x)
                x = nn.relu(x)
                x = nn.avg_pool(x, window_shape=(2, 2), strides=(2, 2))
                x = x.reshape((x.shape[0], -1))  # Flatten the input
                x = nn.Dense(features=64)(x)
                x = nn.relu(x)
                x = nn.Dense(features=10)(x)
                return x

        self.model = SimpleModel()
        self.data = jax.random.normal(jax.random.PRNGKey(0), (100, 28, 28, 1))
        self.labels = jax.random.randint(jax.random.PRNGKey(1), (100,), 0, 10)

    def test_quantize_model(self):
        quantized_model = self.edge_ai.quantize_model(self.model)
        self.assertIsNotNone(quantized_model)
        # Add more specific tests for quantization

    def test_prune_model(self):
        pruned_model = self.edge_ai.prune_model(self.model)
        self.assertIsNotNone(pruned_model)
        # Add more specific tests for pruning

    def test_knowledge_distillation(self):
        student_model = self.edge_ai.knowledge_distillation(self.model, self.model, self.data)
        self.assertIsNotNone(student_model)
        # Add more specific tests for knowledge distillation

    def test_setup_federated_learning(self):
        self.edge_ai.setup_federated_learning(num_clients=5)
        # Add assertions to check if federated learning is set up correctly

    def test_integrate_snn(self):
        snn_model = self.edge_ai.integrate_snn(input_size=784, num_neurons=100)
        self.assertIsNotNone(snn_model)
        # Add more specific tests for SNN integration

    def test_self_healing_mechanism(self):
        healed_model = self.edge_ai.self_healing_mechanism(self.model, self.data)
        self.assertIsNotNone(healed_model)
        # Add more specific tests for self-healing mechanism

    def test_optimize_latency_and_power(self):
        optimized_model, latency, power = self.edge_ai.optimize_latency_and_power(self.model)
        self.assertIsNotNone(optimized_model)
        self.assertIsInstance(latency, float)
        self.assertIsInstance(power, float)
        # Add more specific tests for latency and power optimization

    def test_train_model(self):
        self.edge_ai.train_model(self.model, self.data, self.labels)
        self.assertIsNotNone(self.edge_ai.model)
        # Add more specific tests for model training

    def test_infer(self):
        self.edge_ai.train_model(self.model, self.data, self.labels)
        result = self.edge_ai.infer(jax.random.normal(jax.random.PRNGKey(2), (1, 28, 28, 1)))
        self.assertIsNotNone(result)
        self.assertEqual(result.shape, (1, 10))  # Assuming 10 output classes

if __name__ == '__main__':
    unittest.main()
